fix(app-about): resolve null background image
Setting a background image in the .scss file is not currently working so we had to do a hack-around
throughout the application in order to have the background image properly show (i.e targetting the
element and setting it in the javascript). Will submit a ticket to stenciljs team to see if this is
a current issue.